Adding positive unit tests for Lexer

+++ CParser/Lexing/FloatingToken.cs
@@ -1,9 +1,9 @@
 namespace CParser.Lexing
 {
-    public class FloatingToken : ValueToken<decimal>
+    public class FloatingToken : ValueToken<double>
     {
         public bool HasFloatSpecifier { get; }
-        public FloatingToken(int line, int column, string filename, decimal value, bool hasFloatSpecifier)
+        public FloatingToken(int line, int column, string filename, double value, bool hasFloatSpecifier)
             : base(Terminal.FloatingConstant, line, column, filename, value)
         {
             HasFloatSpecifier = hasFloatSpecifier;
